requirements.txt
openai-whisper==20231117
pytube==15.0.0